Deleted useless assets & Fixed code
I deleted all the unused player characters that were left in the game. There were exactly five copies, each with its own textures, bones, materials, and animations.

In the code, I had to remove the only two public GameObject variables that referenced the players in order to relocate them to another part of the game. I also removed a public Vector3[] that was used to move their positions to the new location. This new location is where the deposits are used for the minigame when arriving at the banks.